rev-list: add option --show-forkpoints
This option makes history simplification keep a commit when it is a parent
commit for multiple relevant commits. The commit in question is where the
project history has diverged into multiple chains, all of which contain
commits interesting to us, even if the commit in question is uninteresting
or treesame.

Signed-off-by: FriendlyNeighborhoodShane <shane.880088.supw@gmail.com>